Move InfoListItemMolecule to the molecules package.
